requirements.txt
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=0.19.0
tqdm>=4.64.0
openpyxl>=3.0.10